fix: fixes in derived resources
